+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral Changelog
 
+## 0.15.13 (tbd)
+
+- Fixed `pilet publish --fresh` with schema defined by the "pilet.json"
+- Fixed duplicated rendering with nested extensions in `piral-blazor` (#602)
+- Added support for `events` capability in `piral-blazor`
+
 ## 0.15.12 (May 19, 2023)
 
 - Fixed the `piral-cli` generating wrong d.ts files for TypeScript 5